fix a bug in kv_vector: Op::ASSIGN -> Op::PLUS

+++ src/filter/frequency_filter.h
@@ -49,7 +49,7 @@ SArray<K> FreqencyFilter<K,V>::QueryKeys(const SArray<K>& key, int freqency) {
   CHECK_LT(freqency, kuint8max) << "change to uint16 or uint32...";
   SArray<K> filtered_key;
   for (auto k : key) {
-    if (count_.query(k) > freqency) {
+    if ((int)count_.query(k) > freqency) {
      filtered_key.push_back(k);
     }
   }
@@ -64,7 +64,7 @@ void FreqencyFilter<K,V>::InsertKeys(const SArray<K>& key, const SArray<V>& coun
   }
 }
 
-// hash implementation
+// DEPRECATED hash implementation
 // std::unordered_map<K, V> map_;
 
 // template<typename K>
